+++ fotoobo/cli/fgt/config.py
@@ -70,7 +70,7 @@ def get(
     ),
     path: str = typer.Argument("/", help="Configuration path", metavar="[path]"),
 ) -> None:
-    """Get configuration or parts of it from one or more FortiGate configuration files"""
+    """Get configuration or parts of it from one or more FortiGate configuration files."""
     result = fgt.config.get(configuration, scope, path)
     result.print_raw()