+++ fotoobo/tools/fgt/cmdb/__init__.py
@@ -1,0 +1,3 @@
+"""
+fgt tools cmdb
+"""