+++ tests/tools/fgt/get/test_api_get.py
@@ -1,0 +1,18 @@
+"""
+Test fgt tools api_get
+"""
+
+# pylint: disable=no-member
+# mypy: disable-error-code=attr-defined
+from pathlib import Path
+
+import pytest
+from _pytest.monkeypatch import MonkeyPatch
+
+
+@pytest.fixture(autouse=True)
+def inventory_file(monkeypatch: MonkeyPatch) -> None:
+    """Change inventory file in config to test inventory"""
+    monkeypatch.setattr(
+        "fotoobo.helpers.config.config.inventory_file", Path("tests/data/inventory.yaml")
+    )