+++ WHATSNEW.md
@@ -1,8 +1,15 @@
 ### Added
 
+- Add generic api_get() method in tools/fgt/get
+- Add CLI command fgt get cmdb firewall address
+- Add CLI command fgt get cmdb firewall addrgrp
+- Add CLI command fgt get cmdb firewall service-custom
+- Add CLI command fgt get cmdb firewall service-group
+- Add the layers of fotoobo into the architecture documentation
+
 ### Changed
 
 - Upgrade pygount to 1.8.0 due to import error
+- Optimize imports in CLI module
 
 ### Removed
-